requirements.txt
fastapi==0.115.12
google-genai==1.47.0
pydantic==2.11.3
pydantic-settings==2.8.1
structlog==25.1.0
googleapis-common-protos==1.70.0
uvicorn[standard]==0.34.0
google-api-core==2.30.0